[REFACTOR] split tools.hpp in individual files
though they are not loosely coupled enough yet to get included directly, but its a step
