+++ inexor/engine/console.cpp
@@ -4,6 +4,7 @@
 #include "inexor/engine/engine.hpp"
 #include "inexor/ui/input/InputRouter.hpp"
 #include "inexor/util/Logging.hpp"
+#include "inexor/shared/cube_queue.hpp"
 
 
 using namespace inexor::util;
