+++ inexor/engine/pvs.cpp
@@ -1,6 +1,7 @@
 #include "inexor/engine/engine.hpp"
 #include "inexor/ui/input/InputRouter.hpp"
 #include "inexor/util/Logging.hpp"
+#include "inexor/shared/cube_queue.hpp"
 
 using namespace inexor::io;
