+++ inexor/shared/cube_loops.hpp
@@ -1,0 +1,24 @@
+/// Legacy file defining loops. Do not use in new code.
+///
+/// (Use normal for loops: for(int i = 0; i < MAX_VALUE; i++)
+///  possibly with ranges: for(auto &iterator : texture_vector) //texture_vector is an std::vector
+#pragma once
+
+#define loop(v,m) for(int v = 0; v < int(m); ++v)
+#define loopi(m) loop(i,m)
+#define loopj(m) loop(j,m)
+#define loopk(m) loop(k,m)
+#define loopl(m) loop(l,m)
+
+// Reverse loops:
+#define looprev(v,m) for(int v = int(m); --v >= 0;)
+#define loopirev(m) looprev(i,m)
+#define loopjrev(m) looprev(j,m)
+#define loopkrev(m) looprev(k,m)
+#define looplrev(m) looprev(l,m)
+
+// macros for looping though vectors
+#define loopv(v)    for(int i = 0; i<(v).length(); i++)
+#define loopvj(v)   for(int j = 0; j<(v).length(); j++)
+#define loopvk(v)   for(int k = 0; k<(v).length(); k++)
+#define loopvrev(v) for(int i = (v).length()-1; i>=0; i--)
